+++ tests/Integration/TestClientFactory.php
@@ -6,6 +6,7 @@ use Http\Adapter\Guzzle7\Client as GuzzlePsrClient;
 use Nyholm\Psr7\Factory\Psr17Factory;
 use Solarium\Client;
 use Solarium\Core\Client\Adapter\Curl;
+use Solarium\Core\Client\Adapter\Http;
 use Solarium\Core\Client\Adapter\Psr18Adapter;
 use Symfony\Component\EventDispatcher\EventDispatcher;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
@@ -27,6 +28,15 @@ final class TestClientFactory
     {
         return new Client(
             new Curl(),
+            $eventDispatcher ?? new EventDispatcher(),
+            $options
+        );
+    }
+
+    public static function createWithHttpAdapter(array $options = null, EventDispatcherInterface $eventDispatcher = null): Client
+    {
+        return new Client(
+            new Http(),
             $eventDispatcher ?? new EventDispatcher(),
             $options
         );